requirements.txt
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Compress==1.4.0
Flask-Cors==3.0.3
Flask-HTTPAuth==3.2.3
Flask-SQLAlchemy==2.2
gevent==1.2.2
greenlet==0.4.13
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
joblib==0.11
jsonify==0.5
MarkupSafe==1.0
MonthDelta==0.9.1
numpy==1.14.1
pandas==0.22.0
passlib==1.7.1
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.11
urllib3==1.22
websocket==0.2.1
websocket-client==0.47.0
Werkzeug==0.12.2
